wc: Factorize priority rank map into web utils method, #TASK-6867
